requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
praw==7.7.1
prawcore==2.3.0
python-lorem==1.3.0.post1
requests==2.31.0
update-checker==0.18.0
urllib3==2.0.5
websocket-client==1.6.3
